refactor: Change return type of method Utils.max
Changed from double to int. Updated dependencies.

+++ src/recognition/Main.java
@@ -79,7 +79,7 @@ public class Main {
         return true;
     }
 
-    private void output(double result) {
-        System.out.println("This number is " + (int) result);
+    private void output(int result) {
+        System.out.println("This number is " + result);
     }
 }
